+++ packages/deepyr/lib/src/base/component_stylings.dart
@@ -87,6 +87,7 @@ mixin AllComponentStylings
         CountdownStyling,
         DiffStyling,
         DividerStyling,
+        DockStyling,
         DrawerStyling,
         DropdownStyling,
         DropdownContentStyling,